+++ custom.d.ts
