+++ src/assets/index.js
@@ -1,13 +1,9 @@
-import linkIcon from './link.svg'
-import loader from './loader.svg'
-import copy from './copy.svg'
-import logo from './logo.svg'
-import tick from './tick.svg'
+// TODO: swap in react-icons instead of svg?
 
-export {
-    linkIcon,
-    loader,
-    copy,
-    logo,
-    tick
-}
+import linkIcon from "./link.svg";
+import loader from "./loader.svg";
+import copy from "./copy.svg";
+import logo from "./logo.svg";
+import tick from "./tick.svg";
+
+export { linkIcon, loader, copy, logo, tick };
