+++ src/pages/Layout/Layout.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react";
-import Header from "../Header";
+import Header from "../../components/Header/Header";
 
 interface LayoutProps {
   children: ReactNode;
