+++ src/pages/PatientManager/PatientTypes.ts
@@ -1,0 +1,9 @@
+export interface PatientInfo {
+  ID: string;
+  Diagnosis: string;
+  OtherDiagnosis: string;
+  Description: string;
+  Age: number;
+}
+
+
